stf to its own crate
