+++ enclave/stf/Cargo.toml
@@ -1,0 +1,94 @@
+[package]
+name = "substratee-stf"
+version = "0.1.0"
+authors = ["Supercomputing Systems AG <info@scs.ch>"]
+edition = "2018"
+
+[features]
+default = []
+sgx = [
+    "sgx_tstd",
+    "env_logger",
+    "log",
+    ]
+
+[dependencies]
+log               = { version = "0.4.7", git = "https://github.com/mesalock-linux/log-sgx", optional=true }
+env_logger        = { version = "0.6.2", git = "https://github.com/mesalock-linux/env_logger-sgx", optional=true }
+
+[dependencies.sgx_tstd]
+rev = "v1.0.9"
+git = "https://github.com/baidu/rust-sgx-sdk"
+features = ["untrusted_fs","net","backtrace"]
+optional = true
+
+[dependencies.codec]
+package = "parity-scale-codec"
+version = "1.0.0"
+default-features = false
+features = ["derive"]
+
+[dependencies.primitives]
+git = "https://github.com/scs/substrate-api-client"
+package = "substrate-primitives"
+branch = "no_std_rstd"
+default-features = false
+
+[dependencies.runtime_primitives]
+git = 'https://github.com/paritytech/substrate'
+rev = "9b08e7ff938a45dbec7fcdb854063202e2b0cb48"
+package = "sr-primitives"
+default-features = false
+
+[dependencies.balances]
+default-features = false
+git = "https://github.com/paritytech/substrate"
+package = "srml-balances"
+rev = "9b08e7ff938a45dbec7fcdb854063202e2b0cb48"
+
+[dependencies.timestamp]
+default-features = false
+git = "https://github.com/paritytech/substrate"
+package = "srml-timestamp"
+rev = "9b08e7ff938a45dbec7fcdb854063202e2b0cb48"
+
+[dependencies.system]
+default-features = false
+git = "https://github.com/paritytech/substrate"
+package = "srml-system"
+rev = "9b08e7ff938a45dbec7fcdb854063202e2b0cb48"
+
+[dependencies.srml-support]
+default-features = false
+git = "https://github.com/paritytech/substrate"
+package = "srml-support"
+rev = "9b08e7ff938a45dbec7fcdb854063202e2b0cb48"
+
+[dependencies.indices]
+default-features = false
+git = "https://github.com/paritytech/substrate"
+package = "srml-indices"
+rev = "9b08e7ff938a45dbec7fcdb854063202e2b0cb48"
+
+[dependencies.version]
+default-features = false
+git = "https://github.com/paritytech/substrate"
+package = "sr-version"
+rev = "9b08e7ff938a45dbec7fcdb854063202e2b0cb48"
+
+[dependencies.sr-io]
+git = "https://github.com/scs/substraTEE-node"
+package = 'sr-io'
+branch = "brenzi-private-tx"
+default-features=false
+features = ["no_oom", "no_panic_handler", "sgx", "debug"]
+
+
+[patch.'https://github.com/paritytech/substrate']
+sr-io = { git = "https://github.com/scs/substraTEE-node", branch = "brenzi-private-tx"}
+substrate-primitives = { git = "https://github.com/scs/substrate-api-client", branch = "no_std_rstd"}
+substrate-application-crypto = { path = "../../application-crypto"}
+srml-balances = { path = "../balances" }
+
+[patch.crates-io]
+ed25519-dalek = { git = "https://github.com/scs/ed25519-dalek.git", branch = "no_std_sgx"}
